% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batsch-package.R
\docType{package}
\name{batsch-package}
\alias{batsch-package}
\alias{_PACKAGE}
\title{batsch: Real-Time PCR Data Sets by 'Batsch et al.' (2008)}
\description{
Real-time quantitative 'polymerase' chain reaction ('qPCR') data sets by 'Batsch et al.' (2008) \doi{10.1186/1471-2105-9-95}. This package provides five data sets, one for each 'PCR' target: (i) rat 'SLC6A14', (ii) human 'SLC22A13', (iii) pig 'EMT', (iv) chicken 'ETT', and (v) human 'GAPDH'. Each data set comprises a five-point, four-fold dilution series. For each concentration there are three replicates. Each amplification curve is 45 cycles long. Original raw data file: \url{https://static-content.springer.com/esm/art\%3A10.1186\%2F1471-2105-9-95/MediaObjects/12859_2007_2080_MOESM5_ESM.xls}.
}
\author{
\strong{Maintainer}: Ramiro Magno \email{ramiro.magno@gmail.com} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Other contributors:
\itemize{
  \item Pattern Institute [copyright holder, funder]
}

}
\keyword{internal}
